Shop Functional
Can now use shop to upgrade stats. Bug occurs when don't have enough
gold and click "back." Wrong gold value expressed. Also message occurs
after cannot buy. Should not be there.

+++ PirateGame/PirateGame/Ship.cs
@@ -44,7 +44,7 @@ namespace PirateGame
             return gold;
         }
 
-        public float getCrew()
+        public int getCrew()
         {
             return crew;
         }
